Include external CSS in SWC lessons
These lessons reference CSS outside of their own directories within the
website, so require spanning across hosts. Thi is limited to the lesson
site to avoid major expansions of the volume of data stored.

This is only done for Software Carpentry because it is not needed for
Data Carpentry and because DC lessons are posted to the organizations
main domain adding -H increases storage by 10x.

Co-authored-by: Virnaliz Cruz <virnaliz.cruz@weecology.org>