Use current R version for miniCRAN repo
This code previously used the system version of R to create the miniCRAN repo. We want
the miniCRAN repo to match that of the R installer, so we now use the same version
using find_r_current_version().

This also fixes an issue where the system version of R pre-dated the provision of two
sets of macOS architectures, resulting in the miniCRAN install failing due to the
different url structures.