Use fixed Python version for Python/PyPI installs
It is increasingly common for packages in the Python data space to have
lagging support for the newest Python release. This means that we could
periodically break due to new Python releases (see #143) and if a
release happened right before a workshop that would be an issue.
Therefore the safer approach with Python is to fix the version and then
periodically test and update after new releases.

This change sets the version to 3.12 and uses this version for all
Python related installation.